pyproject.toml
[project]
name = "vision-cifar-100"
version = "0.1.0"
description = ""
authors = [
    {name = "sebnm"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["torch (==2.2.2)", "torchvision (==0.17.2)", "torchaudio (==2.2.2)", "numpy (==1.26.4)", "pandas (==2.1.2)", "matplotlib (>=3.10.8,<4.0.0)", "seaborn (>=0.13.2,<0.14.0)", "tqdm (>=4.67.3,<5.0.0)", "scikit-learn (>=1.8.0,<2.0.0)", "pyyaml (>=6.0.3,<7.0.0)", "torchinfo (>=1.8.0,<2.0.0)", "torchmetrics (>=1.8.2,<2.0.0)", "wandb (>=0.24.2,<0.25.0)"]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"



[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}
torchaudio = {source = "pytorch"}
[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
